General: Remove "DoubleFormatter" references from Lekce 09
First of all, IMO there isn't all that compelling reason to talk about
locale-aware conversion between numbers and strings (all we gain is the
possibility to use ',' as a decimal point in Czech applications which I
don't believe to be important in this exercise). At the same time, it
may be confusing for the participants since we haven't talked about
anything similar during the lecture(s).

Secondly, the class name is wrong, it's probably supposed to be
DecimalFormat.

Thirdly, DecimalFormat.parse doesn't return double (or Double), it
returns Number. That means that the whole conversion needs to be:
double cislo1 = Double.valueOf(prevodnik.parse(text1));
which adds another static method call to the overall confusion.

Since we use just Integer.parseInt and Integer.toString to deal with
integers, I suggest using Double.parseDouble and Double.toString for
doubles and be done with it.